perf(loaders): honest per-resource readiness flags + zero-copy rkyv access for vocabulary/furigana
Slice 1 (observability):
- routes.rs: each stage-2/3 readiness flag now flips when ITS loader
  finishes (tracked wrapper) instead of after the stage-wide join!
  barrier — the loading overlay counter no longer stalls on '1 of 9'
  for the whole stage-2 wall time. Stage barriers, the solo stage-1
  dictionary load and the terminal is_jlpt_content_loaded signal (set
  after seeding/precache phases) are preserved unchanged.
- Loaders log fetch/cache-read/parse split timings.

Slice 2a (zero-copy access):
- vocabulary module split into {mod, info, chunks, blob}; CDN rkyv blob
  and the local rkyv cache are now consumed via rkyv::access over an
  immortal AlignedVec copy (no owned-structure build, no alloc storm);
  JSON-chunk fallback keeps the owned path behind the same free-function
  lookup API.
- furigana module split into {mod, parse}; FuriganaStore dispatches
  lookups over the owned dictionary (text fallback) and the archived
  view (CDN blob); annotator boundary works on owned entry clones.
- kanji.rs popular_words uses the free lookup function.
- now_ms() is platform-safe on non-wasm targets (tests).

Measured on real CDN data (native): vocab from_bytes 147ms -> access
74ms; furigana 97ms -> 23ms; parity covered by rstest suites.
